allow disabling cleanup

+++ tests_integration/conftest.py
@@ -55,7 +55,9 @@ def random_name_generator(ci_run_number):
 @pytest.fixture
 def run_kopf(kopf_runner_args, kopf_settings):
     @contextmanager
-    def inner(*, enable_connector_reconciler=True, enable_group_reconciler=True):
+    def inner(
+        *, enable_connector_reconciler=True, enable_group_reconciler=True, cleanup=True
+    ):
         env = {}
         if enable_connector_reconciler:
             env["CONNECTOR_RECONCILER_INTERVAL"] = "1"
@@ -73,8 +75,9 @@ def run_kopf(kopf_runner_args, kopf_settings):
             try:
                 yield runner
             finally:
-                kubectl("delete twingate --all")
-                time.sleep(5)
+                if cleanup:
+                    kubectl("delete twingate --all")
+                    time.sleep(5)
 
         assert runner.exception is None
         assert runner.exit_code == 0
